Session 14: decimal quantities + Firestore collectionGroup rules
- quantity fields changed from Int to Double across all request models
  (MaterialToolRequest, MaterialToolPurchase, Transfer) so users can
  enter fractional amounts like 2.5 kg
- inputType changed to numberDecimal in all item row layouts
- fromMap() updated to parse quantity as Number→Double (handles both
  Long and Double stored in Firestore)
- Firestore rules: added {path=**} collectionGroup rules for all 6
  sub-collections so admin portal cross-user reads are not blocked

+++ app/src/main/java/com/raghav/whitecoffee/ui/requests/MaterialToolBuyFragment.kt
@@ -142,7 +142,7 @@ class MaterialToolBuyFragment : BaseFragment<FragmentMaterialToolBuyBinding>() {
         val unit  = row.etUnit.text?.toString()?.trim() ?: ""
         val price = row.etPrice.text?.toString()?.toDoubleOrNull() ?: 0.0
         if (name.isBlank() || qty <= 0) return@mapNotNull null
-        PurchaseItem(name, qty.toInt(), unit, price, qty * price)
+        PurchaseItem(name, qty, unit, price, qty * price)
     }
 
     private fun showSuccessAndExit() {
